+++ app/src/main/java/com/africanbongo/whipitkotlin/storage/database/model/DatabaseIngredient.kt
@@ -8,9 +8,9 @@ import zw.co.bitpirates.spoonacularclient.model.Ingredient
  * Mapping of an [Ingredient] that will be stored in the database.
  */
 @Entity(tableName = "ingredient_table")
-data class DatabaseIngredient (
+data class DatabaseIngredient(
     @PrimaryKey(autoGenerate = false) val ingredientId: Int,
     val description: String,
-    val amount: Float,
+    val amount: Double,
     val unit: String?
 )
