+++ app/src/main/java/com/africanbongo/whipitkotlin/storage/database/model/DatabaseRecipe.kt
@@ -7,13 +7,13 @@ import androidx.room.PrimaryKey
  * Recipe class that's used to store and retrieve objects in the database.
  */
 @Entity(tableName = "recipe_table")
-data class DatabaseRecipe (
+data class DatabaseRecipe(
    @PrimaryKey val recipeId: Int,
    val title: String,
    val imageUrl: String?,
    val sourceName: String,
    val sourceUrl: String,
-   val spoonacularScore: Float,
+   val spoonacularScore: Double,
    val servings: Int,
    val readyInMinutes: Int,
 )
